requirements.txt
requests>=2.13.0
attrs>=18.1.0
opentracing>=1.3.0
uvloop>=0.10.1
